+++ tests/CustomerMandateTest.php
@@ -226,6 +226,33 @@ class CustomerMandateTest extends ResourceTestCase
     }
 
     /**
+     * Revoke a mandate
+     */
+    public function testRevokeCustomerMandate()
+    {
+        // Mock the mandate
+        $mandateMock = $this->getMandate();
+
+        // Mock the customer
+        $customerMock = $this->getCustomer();
+
+        // Mock the request
+        $requestMock = $this->createMock(Request::class);
+
+        $requestMock
+            ->expects($this->once())
+            ->method('delete')
+            ->with($this->equalTo("/customers/{$customerMock->id}/mandates/{$mandateMock->id}"));
+
+        // Create API instance
+        $api = new Mollie('test_testapikey');
+        $api->request = $requestMock;
+
+        // Revoke mandate
+        $api->customer($customerMock->id)->mandate($mandateMock->id)->revoke();
+    }
+
+    /**
      * Create customer mandate with invalid signature date
      *
      * @expectedException InvalidArgumentException